Refactored auth to auth route
