+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon server/server-spec.js",
+    "start": "nodemon server/server.js",
     "test": "mocha test/**/*spec.js --exit",
     "pack": "webpack --watch",
     "dev": "node node_modules/.bin/webpack-dev-server --history-api-fallback --hot --inline --progress"
